add remaining days in wipe to btc profit calc
